Resolve full path for cirrus version
The versions of Cirrus is given by major.minor.patch-build+hash, and
less granular version as symlinked (e.g. 1.9 may be linked to
1.9.1-1.+c26718b)

Realizations do not always run in parallell, and we could end up in a
situation where two different builds are used. Consider that the user
specifies version 1.9, and for the first realization this resolves in
one versions, but for a subsequent realization a new build of cirrus has
been made which resolves in a slightly different version.

We want to make sure that the exact same version of Cirrus is used
throughout an Ert run, and therefore handle the resolvement upfront.