Remove suggestion to specify num cpu
By default, as we use <NUM_CPU> in the command, the NUM_CPU from the ert
user config will be used. It is not necessary for the user to also
specify it.

This does not stop the user from adding it, hence overriding the NUM cpu
from Ert is still possible.